Fix songs that are entirely ALL CAPS.
Lyrics in ALL CAPS is considered as shouting and therefore discouraged/bad style. Thus, if a song’s lyrics are entirely (!) ALL CAPS, they are all changed to lowercase and then the first word (more precisely the first capitalizable letter of the first word) of each line is capitalized again.

The only downside is that names/nouns that should be capitalized are then lowercase, but this should only concern a handful of words, so fixing this manually only takes a small effort.