Revert "Added io error handling."
IOError doesn't necessarily indicate a permisison problem. E.g. it also includes
FileNotFoundError, which in one case was caused by yt-dlp not being able to
locate the cookie file.
Also, catching all IOErrors (OSErrors) makes debugging harder.
If we want to handle permission issues separately, we should instead handle
the error when actually doing the file op, so we can be sure that's really the issue.

This reverts commit c03f32212b322cd8fa0b932512fc215f532c6b21.