Proof-of-concept of indicating image quality in the song list.
This reads out the image resolution of cover and background files and displays the quality according to some prefined thresholds via colored check marks in the respective song table columns. If the cover images are non-square or the background images are non-16:9-widescreen format, this is additionally indicated by a small exclamation mark on the check mark icon.

While this works as expected, it is highly inefficient because the image files are read on the fly. It would probably be better to store the image dimensions in the meta file and simply use that information.

The inefficiency is doubled as the image dimensions are determined twice, once to determine the icon to display, a second time to determine the appropriate tooltip.