Added some comments to clarify audio processing.
I realized that this branch also fixes a regression that is already in main. Selecting an audio format other than m4a would lead to a crash because prepare_filename() only returns the filename before any postprocessing. It's getting a little complicated with target codecs, postprocessors and normalization, that's why I thought it would be helpful to add some explanatory comments here.