Do not delete #VIDEOGAP, but keep warning.
There are a few exceptions where the video does not align to its own audio track, so a #VIDEOGAP fixes async audio and video of the same source. Therefore, we keep the #VIDEOGAP even if audio and video stem from the same source, but also keep the warning, because it the described situation is rare and the usage of #VIDEOGAP is most likely not intended or the resources are incomplete.