Refactor song_loader.
This adds a new class Job and a function register_job() with a decorator so that all relevant functions that need to be called for each song can be labeled accordingly. It also adds return the class JobResult as return value class to allow for improved error handling and user feedback.