Auto-correct wrong medley sections.
In case the medley start beat does not coincide with the first beat of the first note of the line, it is shifted to that. Similarly, if the medley end beat does not coincide with the last beat of the last note of the line, it is shifted there. Furthermore, warnings for unusually short (< 20 seconds) or unusually long (> 80 seconds) medley sections are logged.