Linux bundle testing improvements (#496)
This is supposed to reduce friction developing and testing linux bundles.

The core idea is a docker image that is built manually and rarely changes. It comes with python built from source. When a bundle is needed, this docker image is used to build it.

In addition to heavily reducing release time (because pulling the image is infinitely quicker than building python), this allows for reasonably quick development builds on any branch (https://docs.github.com/en/actions/how-tos/manage-workflow-runs/manually-run-a-workflow). So when a problem such as #495 occurs, one could easily attempt to fix the issue, run the new linux-build workflow on that branch and only needs to check the fix on the affected system (in that case, Debian 13), which can be solved also by asking the affected person to try out the new build.

Control over the build image also means it is fairly easy to change and test the build environment in case a new dependency is needed.